+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 
 - Fixed infinite tight loop in PVOutput status service when lock times out
+- Fixed "Task was destroyed but it is pending!" warning during restart after firmware upgrade (or other restart event)
 
 ---